8. FORMATTING - moved routes to routes folders
