+++ Develop/db/db.json
@@ -1,1 +1,1 @@
-[{"title":"Test Title","text":"Test text"},{"title":"dasdasdasdas","text":"asdasdasdasd","uuid":"c80b8b9e-ae3e-4fc1-a1ec-dbfc9307f627"},{"title":"asdasdasdasd","text":"asdasdasdasdasdasdasd","uuid":"3a6d0630-1cbb-4253-86bb-8814dc6f2d45"},{"title":"fjwjifwjpwdjpc","text":"fhnfdnfqpwdjcnqwdcqwdc","uuid":"6ebf503f-b18e-4712-ad89-3739c2a2f625"}]
+[]
