+++ Develop/routes/html-routes.js
@@ -1,10 +1,13 @@
 const path = require ("path");
+const htmlRoutes = require("express").Router();
 
 
-app.get("/notes", (req, res) => {
-    res.sendFile(path.join(__dirname, "public/notes.html"))
-})
-
-app.get("*", (req, res)=>{
-    res.sendFile(path.join(__dirname, "public/index.html"))
+htmlRoutes.get("/notes", (req, res) => {
+    res.sendFile(path.join(__dirname, "../public/notes.html"))
 });
+
+htmlRoutes.get("*", (req, res)=>{
+    res.sendFile(path.join(__dirname, "../public/index.html"))
+});
+
+module.exports = htmlRoutes
